requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
Markdown==3.6
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.2
